+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for namedns
 
+## Upcoming release
+
+- Add `--api-url` and `--verbose` flags
+
 ## 1.2
 
 - Add [XDG](https://standards.freedesktop.org/basedir-spec/basedir-spec-latest.html) config directory support